SJ-Add: Agregando endpoints faltantes de entidades necesarias

+++ SMJRegisterAPI/Entities/Camper.cs
@@ -8,7 +8,7 @@ public class Camper : BaseEntity
     public string Name { get; set; }
     public string LastName { get; set; }
     public string DocumentNumber { get; set; }
-    public int  PaidAmount { get; set; }
+    public decimal  PaidAmount { get; set; }
     public bool IsGrant { get; set; } = false;
     public bool IsPaid { get; set; } = false;
     public Gender Gender { get; set; }
